Remove pointer non-null checks before free()
Change ced33db764ff35e0bd35754391814e29602f4233 reported a bug that
was fixed by adding a check to verify a pointer is not null before
freeing.  POSIX defines free() as performing no action when passed a
NULL pointer.

It's a bug that they're inconsistent and this is inconsistent with
what `free(p)` is designed to do.

If this bug actually exists somewhere, we need a test case to
reproduce it, but it doesn't seem to be this.