Buffer file I/O
Implemented a new set of ops in iobuffer.c/h that call through to the default ops,
but buffer the data.
There's a single write buffer that collects data being written and writes it to the file when
it fills up, or the file position changes, or sync or close are called.
There's a set of read buffers that cache regions of the file on an LRU basis.

Altogether these about double the speed of testapp.c.

Change-Id: I0ac74e997d5b00d576bd37ed0ea2f6e554a0e32a
Reviewed-on: http://review.couchbase.org/14866
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Damien Katz <damien@couchbase.com>