Use arena memory allocator during b-tree modification
modify_btree() allocates a lot of small blocks and then frees them all
(except the root) at the end. We can reduce the overhead of allocation
by using a simple arena allocator, which allocates by just bumping a
pointer and doesn't bother to free blocks individually, just when the
arena itself is freed.

Change-Id: I44540003bf813ee26ed93d5d340ddf36b03f4ca9
Reviewed-on: http://review.couchbase.org/14879
Tested-by: Aaron Miller <apage43@ninjawhale.com>
Reviewed-by: Aaron Miller <apage43@ninjawhale.com>