Better bit and packed-struct twiddling.
Replaced the bit-access functions in bitfield.h with a series of
structs representing big-endian integers of sizes 1-6 bytes, and
a set of functions to read and write them.

This then allows us to define actual structs for the database data
structures; these are in node_types.h.

Then use those structs when reading and writing structured data,
which makes the code a lot clearer since it's accessing named fields
instead of magic offset/size constants.

Change-Id: I7266637d2d82155b21339220a8e0c2b1f5f18516
Reviewed-on: http://review.couchbase.org/15559
Reviewed-by: Jens Alfke <jens@couchbase.com>
Tested-by: Jens Alfke <jens@couchbase.com>