MB-6293: Better bit and packed-struct twiddling.
(Revised to obey C strict aliasing rules!)

Replaced the bit-access functions in bitfield.h with a series of
structs representing big-endian integers of sizes 1-6 bytes, and
a set of functions to read and write them.

This then allows us to define actual structs for the database data
structures; these are in node_types.h.

Then use those structs when reading and writing structured data,
which makes the code a lot clearer since it's accessing named fields
instead of magic offset/size constants.

Change-Id: I4addff9d215f6f6cfe67f86442dcf70c85a47d60
Reviewed-on: http://review.couchbase.org/19815
Reviewed-by: Aaron Miller <apage43@ninjawhale.com>
Tested-by: Aaron Miller <apage43@ninjawhale.com>