Fix: btree_lookup skips items when called with multiple keys
An off-by-one bug in btree_lookup_inner meant that, when called with
multiple keys, it would end up skipping a key after the end of a btree
node.

This wasn't apparent before, because btree_lookup was never called with
multiple keys until I started working on couchstore_docinfos_by_sequence.

Change-Id: I59ba81058ec2672c88bcdc4aab148872e5f9aa92
Reviewed-on: http://review.couchbase.org/15357
Reviewed-by: Aaron Miller <apage43@ninjawhale.com>
Tested-by: Aaron Miller <apage43@ninjawhale.com>