MB-100 Fix dbdump reading deleted docs
The corrupt node check was checking only the bp field of DocInfos in, as
a zeroed out docinfo is used to indicate b-tree corruption. However, it
is possible for a bp to be 0 in normal cases, being deleted docs.

Also changed the wording of errors reading the doc body, to make the
deleted doc without a body case less scary sounding, as it's not
necessarily an error if we do not expect a body. It is possible for a
deleted doc to have a body, so we do still check for one.

Change-Id: I5e44484eb1cea65496e2ad342157b42ded4ec89d
Reviewed-on: http://review.couchbase.org/19634
Reviewed-by: Jens Alfke <jens@couchbase.com>
Tested-by: Aaron Miller <apage43@ninjawhale.com>