MB-6921: Return NO_HEADERS when opening a non-existent file READ_ONLY
When doing a get on a file that doesn't have headers and we pass in
the READ_ONLY flag from ep-engine and this will cause us to return
the error CHECKSUM_FAILED which is converted into a TMP_FAIL in
ep-engine. It would be much more clear to return a NO_HEADER
error since ew can infer that the file doesn't have any data yet. Then
from the ep-engine side we can map this error to not found.

Change-Id: I058c49e26a2bc6186bff62959a232ab160d4cde4
Reviewed-on: http://review.couchbase.org/21647
Tested-by: Michael Wiederhold <mike@couchbase.com>
Reviewed-by: Aaron Miller <apage43@ninjawhale.com>
Tested-by: Aaron Miller <apage43@ninjawhale.com>
Reviewed-by: Jens Alfke <jens@couchbase.com>