Don't export internal APIs in libcouchstore
It's desired to test some internal APIs, and at the same time
leave them hidden to users of libcouchstore.
Therefore instead of making all these internal APIs exportable
for the sake of allowing the test application to test them,
build a version of libcouchstore named libcouchstore_tests that
exports all internal functions. This library is built exactly
the same way as the regular one, except that it doesn't control
visibility of the internal APIs, allowing them to be used by the
test application.

Change-Id: Ida64e4b44e3b81cae425ba7a842dbcb89e739c86
Reviewed-on: http://review.couchbase.org/25639
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Fulu Li <fulu@couchbase.com>