CBIDXT-255: Support compaction for spatial views
The native compactor in Couchstore now also supports the compaction of
spatial views.

As the spatial views have a similar on-disk format (just key-value pairs)
as mapreduce views have, most of the code can be re-used. The only difference
is the way the tree is built bottom-up. When a parent node is created, the
spatial views need to calculate the bounding box that encloses all children
(as opposed to the mapreduce view where the smallest key is used as key of
the parent node).

Change-Id: I15a1a6ad7f0b69c3610e57d43f1be468dd213da8
Reviewed-on: http://review.couchbase.org/44394
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: buildbot <build@couchbase.com>