MB-13219: Read only files should not create a write buffer
Write buffers are by default created with 128kb of space. This
is unused if a read-only file is opened so we should set the write
buffer to 0. This increases the amount of backfills we can have
running at any given time and has a major affect on memory usage
when using buckets with a small memory quota.

For out minimum bucket size (100mb) this allows us to have about
4096 backfills running at any given time. Without this change we
can only have around 300 open before we run out of memory and the
server becomes unusable.

Change-Id: I2ad9aebc8b55f496ca07994d0d419a91f1df6da6
Reviewed-on: http://review.couchbase.org/45916
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>