MB-15049 Fix the initial tree building with large node size
If the KP node size is large (e.g. due to large reduce value),
while building the btree, we may be flushing that node only to
disk and creating one parent node for that.This will not reduce
the number of nodes in the upper level of the tree, and the
tree building will never finish. In this fix, we make sure that
we flush at least 2 KP nodes.

Change-Id: I9832a232c724fc4dd8ff5218b5217b485374d783
Reviewed-on: http://review.couchbase.org/48072
Reviewed-by: Nimish Gupta <nimish@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>
Reviewed-on: http://review.couchbase.org/51852