MB-17420: Fix stderr on Windows
Set the stderr to binary mode and don't buffer it.

In text mode printed line feeds (`\n`) are automatically prefixed with
carriage returns (`\r'), this leads to error messages containing a
carriage return at the end.

The buffering may lead to error messages not being received on the
Erlang side.

Change-Id: I702051f5baa3b7e10b645477d53bc6abf914e1b5
Reviewed-on: http://review.couchbase.org/58846
Reviewed-by: Harsha H S <hhs.couchbase@gmail.com>
Tested-by: buildbot <build@couchbase.com>