Refactor couchstore_commit to make commit testable
This commit splits commit into two phases

- Precommit: Extends the file to be large enough for
the header and performs a sync. This is to write the
data before the header and to avoid doing a second
metdata flush
- Commit: Actually writes the header which points
to the written data.

This is done to make it possible to test that the
precommit has written the file to be large enough
to include the header.

Change-Id: I9d9f9540544bfc08d1f99240149aee3bc8f28262
Reviewed-on: http://review.couchbase.org/59388
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>