Add error injection tests for fileops
This changes adds a large number of tests for the
purposes of testing how Couchstore reacts to errors
which originate from the file system.

It adds a class MockOps, an implementation of
FileOpsInterface, which can be used to inject errors
at the File Ops layer.

This has been used to add a test which injects an
error for every file ops code-path on the public API
to ensure that errors are propagated properly and
immediately.

Change-Id: I7f8197daf8336b485542e5660ba13ee44bc37906
Reviewed-on: http://review.couchbase.org/61023
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>