Remove 100x loops in mapreduce tests
The couchstore testapp currently takes a very long time to run
and occasionally times out after 1500s on commit validation.

This change removes the code to run all the mapreduce tests 100
times and instead runs them just once.

The loop was previously there to reproduce a V8 crash in the
event that a regression occurred. This should be no longer
necessary as we now run the tests under Valgrind which should
detect race conditions and memory erors.

Change-Id: I8a7dc1285a390726dbee7be23938e5fe9125fc1b
Reviewed-on: http://review.couchbase.org/60717
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>