Remove couchstore_reopen_file
This patch removes couchstore_reopen_file as it is completely unused,
does not function properly under windows and the semantics of it are
broken (Everytime a file is reopened it will wrap it in another
iobuffer).

couchstore_drop_file is left as a later patch will make use of it.

Change-Id: I40936ea8d992a353bca3d1ffab27a7d6aaba3671
Reviewed-on: http://review.couchbase.org/61148
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>