Resolve memory leaks in compaction
couch_compact does not currently close the db which results in a
(short-lived) memory leak reported by valgrind.

Additionally if we compact into a target file which already
exists then we will load in the id header and seq header from the
target file and overwrite them without freeing them.

Change-Id: I2f8756e37c6f71b7fe59aa32501df6f76eb53ff3
Reviewed-on: http://review.couchbase.org/61564
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>