MB-19783 Use barrier to prevent exit of long running task
There is a potential race in which terminator thread decides to
terminate map reduce context and gets switched out of context.
In the meantime map reduce task is finished normally and the same
context is reused to perform map reduce for subsequent mutations.
Now terminator thread will potentially terminate this new task,
when it is scheduled back to run.
This is avoided by using mutex in which either terminator thread,
if successful will terminate long running map reduce task or the
task is terminated normally.

Change-Id: I86baa697da17aa0e87dd8fae82452231aaa22f20
Reviewed-on: http://review.couchbase.org/64790
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Volker Mische <volker.mische@gmail.com>