CBCS-4, MB-19995: Make read buffer unit size and count configurable
* We can use the second LSByte of open flags as configurable parameters
for read buffer unit size (first 4 bits) and buffer count (next 4 bits).
Unit size and count can be ranged from 1KB to 16MB and 8 to 128K,
respectively.

* Setting flags to zero represents using the default value.

* To do: current buffer implementation is based on LRU linked list,
which will be inefficient for larger buffer count (> 100). We need to
use an index such as simple hash table in addition to LRU list, to
retrieve the buffer corresponding to the requested offset quickly.

Change-Id: Icb34d41e2fa3869b095032d8abbf83c1756ea4c0
Reviewed-on: http://review.couchbase.org/73471
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>