Add support for per-filehandle statistics
Add a new function to the couchstore public API -
couchstore_get_db_filestats(). This returns an optional FHStats
interface which callers can use to obtain statistics about the
underlying file handle.

The FHStats interface initially supports two methods:

- getReadCount() to return the number of read() calls on this handle.
- getWriteCount() to return the nubmer of write() calls.

These is needed by kv_engine as it wants to obtain statistics on a
per-file basis, as recorded by StatsOps (couch-fs-stats.h).

Note that the included FileOps implementations in couchstore to *not*
suppport this API currently; one would need to use ep-engine's StatOps
to make use of them. However it may be advantageous to move some of
the stats tracking directly into couchstore in future.

Change-Id: Ie1fff97dad9fd5a95e829ff61b91a6ec5f5a975b
Reviewed-on: http://review.couchbase.org/82316
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>